Rename scenario to region
